+++ angular.json
@@ -22,9 +22,6 @@
               "src/assets"
             ],
             "styles": [
-              // "./node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
-              // "src/styles.css"
-
               "src/assets/vendors/bootstrap/css/bootstrap.min.css",
               "src/assets/vendors/metisMenu/metisMenu.min.css",
               "src/assets/dist/css/sb-admin-2.min.css",
@@ -78,8 +75,10 @@
             "tsConfig": "src/tsconfig.spec.json",
             "karmaConfig": "src/karma.conf.js",
             "styles": [
-              // "./node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
-              // "src/styles.css"
+              "src/assets/vendors/bootstrap/css/bootstrap.min.css",
+              "src/assets/vendors/metisMenu/metisMenu.min.css",
+              "src/assets/dist/css/sb-admin-2.min.css",
+              "src/assets/vendors/font-awesome/css/font-awesome.min.css"
             ],
             "scripts": [],
             "assets": [
