--- src/app/app.component.2.html
@@ -1,34 +1,0 @@
-
-<!-- <app-header *ngIf="authenticated() == false"></app-header> -->
-<!-- <app-secure-header *ngIf="authenticated()"></app-secure-header> -->
-<!-- <body> -->
-<div id="wrapper">
-
-<app-header></app-header>
-<!-- <app-sidenav></app-sidenav> -->
-
-<!-- <app-secure-header></app-secure-header> -->
-<!-- <app-secure-nav></app-secure-nav> -->
-
-        <!-- Page Content -->
-        <div id="page-wrapper" style="min-height: 317px;">
-            <div class="container-fluid">
-                <div class="row">
-                    <div class="col-lg-12">
-                        <h1 class="page-header">Blank</h1>
-                        <router-outlet></router-outlet>
-                    </div>
-                    <!-- /.col-lg-12 -->
-                </div>
-                <!-- /.row -->
-            </div>
-            <!-- /.container-fluid -->
-        </div>
-        <!-- /#page-wrapper -->
-
-<!-- <app-footer></app-footer> -->
-
-</div>
-<!-- /#wrapper -->
-
-
