+++ src/app/app.module.ts
@@ -3,15 +3,15 @@ import { NgModule } from '@angular/core';
 import { HttpModule } from '@angular/http';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { AppComponent } from './app.component';
-import { LoginComponent } from './components/login/login.component';
+import { LoginComponent } from './components/public/login/login.component';
 import { ContentAPIService } from './services/content.api.service';
-import { SignupComponent } from './components/signup/signup.component';
+import { SignupComponent } from './components/public/signup/signup.component';
 import { AppRoutingModule } from './app-routing.module';
-import { HeaderComponent } from './components/header/header.component';
-import { FooterComponent } from './components/footer/footer.component';
+import { HeaderComponent } from './components/public/header/header.component';
+import { FooterComponent } from './components/common/footer/footer.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LayoutModule } from '@angular/cdk/layout';
-import { HomeComponent } from './components/home/home.component';
+import { HomeComponent } from './components/public/home/home.component';
 import { HttpErrorHandler } from './services/http-error-handler.service';
 import { MessageService } from './services/message.service';
 import { SecureHomeComponent } from './components/secure/secure-home/secure-home.component';
@@ -19,14 +19,14 @@ import { ProfileComponent } from './components/secure/profile/profile.component'
 import { LoginService } from './services/auth/login.service';
 import { FormsModule } from '@angular/forms';
 import { JwtInterceptor } from './services/jwt.interceptor';
-import { CommonComponent } from './components/common/common.component';
+import { CommonComponent } from './components/common/common/common.component';
 
 
 import { JwtModule   } from '@auth0/angular-jwt';
 import { SecureHeaderComponent } from './components/secure/secure-header/secure-header.component';
-import { SecureNavComponent } from './components/secure/secure-nav/secure-nav.component';
 import { InjectionToken } from '@angular/core';
-import { SidenavComponent } from './components/sidenav/sidenav.component';
+import { PlaceholderComponent } from './components/common/placeholder/placeholder.component';
+
 
 export function tokenGetter() {
   return localStorage.getItem('token');
@@ -48,9 +48,8 @@ export const CONTENT_LIST: InjectionToken<string> = new InjectionToken<string>('
     SecureHomeComponent,
     ProfileComponent,
     SecureHeaderComponent,
-    SecureNavComponent,
     CommonComponent,
-    SidenavComponent
+    PlaceholderComponent
   ],
   imports: [
     FormsModule,
