+++ src/app/components/common/footer/footer.component.ts
