+++ src/app/components/common/placeholder/placeholder.component.html
@@ -1,0 +1,3 @@
+<p>
+  Placeholder Page: {{page}}
+</p>
