--- src/app/components/footer/footer.component.1.css
@@ -1,14 +1,0 @@
-
-.footer {
-    overflow: hidden;
-    background-color: #f1f1f1;
-    text-align: center;
-    align-self: auto;
-    padding: 5px 2px;
-    position: absolute;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    margin: 1em;
- 
-  }
