--- src/app/components/header/header.component.1.html
@@ -1,16 +1,0 @@
-
-<div class="header" *ngIf="authenticated() == false">
-    <a href="" class="logo">SmartWire</a>
-    <div class="header-right">
-        <div id="myDIV">
-            <ng-container *ngFor="let button of buttons; let i = index" >
-                <button 
-                    [ngClass]="{'active': button.active}" 
-                    (click)="toggleClass(button)" 
-                    [routerLink]="[button.path]">
-                    {{ button.name }}
-                </button>
-            </ng-container>
-       </div>
-  </div>
-</div>
