--- src/app/components/secure/secure-header/secure-header.component.1.css
@@ -1,69 +1,0 @@
-
-.header {
-  overflow: hidden;
-  background-color: #f1f1f1;
-  padding: 8px 4px;
-}
-
-.header a {
-  float: left;
-  color: black;
-  text-align: center;
-  padding: 4px;
-  text-decoration: none;
-  font-size: 18px; 
-  line-height: 25px;
-  border-radius: 4px;
-}
-
-.header a.logo {
-  font-size: 25px;
-  font-weight: bold;
-}
-
-.header a:hover {
-  background-color: #ddd;
-  color: black;
-}
-
-.header a.active {
-  background-color: dodgerblue;
-  color: white;
-}
-
-.header-right {
-  float: left;
-  color: black;
-  text-align: center;
-  padding: 4px;
-  text-decoration: none;
-  font-size: 18px; 
-  line-height: 25px;
-  border-radius: 4px;
-  float: right;
-}
-
-@media screen and (max-width: 500px) {
-  .header a {
-    float: none;
-    display: block;
-    text-align: left;
-  }
-  .header-right {
-    float: none;
-  }
-}
-  .btn {
-    border: none;
-    outline: none;
-    padding: 4px 8px;
-    background-color: #f1f1f1;
-    cursor: pointer;
-    font-size: 18px;
-}
-
-/* Style the active class, and buttons on mouse-over */
-.active, .btn:hover {
-    background-color: #666;
-    color: white;
-}
