--- src/app/components/secure/secure-nav/secure-nav.component.1.html
@@ -1,80 +1,0 @@
-<ng-container *ngIf="authenticated()">
-    <nav id="menu">
-      <label for="tm" id="toggle-menu">Navigation <span class="drop-icon">▾</span></label>
-      <input type="checkbox" id="tm">
-      <ul class="main-menu clearfix">
-        <li><a href="#">Account
-          <span class="drop-icon">▾</span>
-          <label title="Toggle Drop-down" class="drop-icon" for="sm3">▾</label>
-        </a>
-        <input type="checkbox" id="sm3">
-        <ul class="sub-menu">
-          <li><a [routerLink]="['secure-home']">Account Overview</a></li>
-          <li><a href="#">Current Bill</a></li>
-          <li><a href="#">Previous Bills</a></li>
-          <li><a href="#">Payments</a></li>
-        </ul>
-      </li>
-      <li><a href="#">Network 
-        <span class="drop-icon">▾</span>
-        <label title="Toggle Drop-down" class="drop-icon" for="sm1">▾</label>
-      </a>
-      <input type="checkbox" id="sm1">
-        <ul class="sub-menu">
-          <li><a href="#">Current Plan</a></li>
-          <li><a href="#">Performance
-              <span class="drop-icon">▾</span>
-              <label title="Toggle Drop-down" class="drop-icon" for="sm2">▾</label>
-            </a>
-            <input type="checkbox" id="sm2">
-            <ul class="sub-menu">
-              <li><a href="#">My Network</a></li>
-              <li><a href="#">Facility Network</a></li>
-              <li><a href="#">Internet</a></li>
-            </ul>
-          </li>
-          <li><a href="#">Outages</a></li>
-          <li><a href="#">Audit Performance</a></li>
-        </ul>
-      </li>
-      <li><a href="#">Services
-        <span class="drop-icon">▾</span>
-            <label title="Toggle Drop-down" class="drop-icon" for="sm1">▾</label>
-          </a>
-          <input type="checkbox" id="sm1">
-          <ul class="sub-menu">
-            <li><a href="#">Internet</a></li>
-            <li><a href="#">Video</a></li> 
-            <li><a href="#">Telephone</a></li>
-            <li><a href="#">Security</a></li>
-            <li><a href="#">Environment</a></li>
-          </ul>
-        </li>
-        <li><a href="#">Profile
-            <span class="drop-icon">▾</span>
-            <label title="Toggle Drop-down" class="drop-icon" for="sm4">▾</label>
-          </a>
-          <input type="checkbox" id="sm4">
-          <ul class="sub-menu">
-            <li ><a [routerLink]="['secure-profile']">Contact Information</a></li>
-            <li><a href="#">Payment Information</a></li>
-            <li><a href="#">Notification Preferences</a></li>
-          </ul>
-        </li>
-        <li><a href="#">Security
-            <span class="drop-icon">▾</span>
-            <label title="Toggle Drop-down" class="drop-icon" for="sm5">▾</label>
-          </a>
-          <input type="checkbox" id="sm5">
-          <ul class="sub-menu">
-            <li><a href="#">Change Password</a></li>
-            <li><a href="#">Change Security Questions</a></li>
-          </ul>
-        </li>
-      </ul> 
-    </nav>
-    <!-- not sure why we need these breaks, need to figure out a different solution -->
-    <br><br><br>
-    </ng-container>
-    
-
