--- src/assets/ie91.css
@@ -1,39 +1,0 @@
-/*
-	Prism by TEMPLATED
-	templated.co @templatedco
-	Released for free under the Creative Commons Attribution 3.0 license (templated.co/license)
-*/
-
-/* Split */
-
-	.split:after {
-		content: '';
-		display: block;
-		clear: both;
-	}
-
-	.split > * {
-		float: left;
-	}
-
-/* Spotlight */
-
-	.spotlight:after {
-		content: '';
-		display: block;
-		clear: both;
-	}
-
-	.spotlight .image {
-		float: left;
-	}
-
-	.spotlight .content {
-		float: left;
-	}
-
-/* Banner */
-
-	#banner {
-		background: #4A4771;
-	}
