--- src/index.1.html
@@ -1,34 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>Smart Wire</title>
-  <base href="/">
-
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  
-  <!-- <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet"> -->
-  <!-- <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet"> -->
-  <!-- <link href="reset.css" rel="stylesheet"> -->
-  <!-- <link href="normalize.css" rel="stylesheet"> -->
-  <!-- <link href="assets/bootstrap.css" rel="stylesheet"> -->
-  <!-- <link href="styles.css" rel="stylesheet"> -->
-  
-</head>
-<body>
-  <app-root></app-root>
-
-<!-- jQuery -->
-<script type="text/javascript" src="./assets/vendors/jquery/jquery.min.js"></script>
-
-<!-- Bootstrap Core JavaScript -->
-<script type="text/javascript" src="./assets/vendor/bootstrap/js/bootstrap.min.js"></script>
-
-<!-- Metis Menu Plugin JavaScript -->
-<script type="text/javascript" src="./assets/vendor/metisMenu/metisMenu.min.js"></script>
-
-<!-- Custom Theme JavaScript -->
-<script type="text/javascript" src="./assets/dist/js/sb-admin-2.min.js"></script>
-
-</body>
-</html>
